+++ androidtv/adb_manager/adb_manager_async.py
@@ -1,40 +1,37 @@
 """Classes to manage ADB connections.
 
 * :py:class:`ADBPython` utilizes a Python implementation of the ADB protocol.
-* :py:class:`ADBServer` utilizes an ADB server to communicate with the device.
 
 """
 
 
-from contextlib import contextmanager
+import asyncio
+from contextlib import asynccontextmanager
 import logging
-import sys
-import threading
 
-from adb_shell.adb_device import AdbDeviceTcp
-from adb_shell.auth.sign_pythonrsa import PythonRSASigner
+from aio_adb_shell.adb_device import AdbDeviceTcp
+from aio_adb_shell.auth.sign_pythonrsa import PythonRSASigner
 from ppadb.client import Client
 
-from .constants import DEFAULT_AUTH_TIMEOUT_S
-from .exceptions import LockNotAcquiredException
+from ..constants import DEFAULT_AUTH_TIMEOUT_S
+from ..exceptions import LockNotAcquiredException
 
 _LOGGER = logging.getLogger(__name__)
 
-#: Use a timeout for the ADB threading lock if it is supported
-LOCK_KWARGS = {'timeout': 3} if sys.version_info[0] > 2 and sys.version_info[1] > 1 else {}
-
-if sys.version_info[0] == 2:  # pragma: no cover
-    FileNotFoundError = IOError  # pylint: disable=redefined-builtin
+#: Default timeout for acquiring the async lock that protects ADB commands
+DEFAULT_TIMEOUT = 3.0
 
 
-@contextmanager
-def _acquire(lock):
-    """Handle acquisition and release of a ``threading.Lock`` object with ``LOCK_KWARGS`` keyword arguments.
+@asynccontextmanager
+async def _acquire(lock, timeout=DEFAULT_TIMEOUT):
+    """Handle acquisition and release of an ``asyncio.Lock`` object with a timeout.
 
     Parameters
     ----------
-    lock : threading.Lock
+    lock : asyncio.Lock
         The lock that we will try to acquire
+    timeout : float
+        The timeout in seconds
 
     Yields
     ------
@@ -48,17 +45,22 @@ def _acquire(lock):
 
     """
     try:
-        acquired = lock.acquire(**LOCK_KWARGS)
-        if not acquired:
+        acquired = False
+        try:
+            acquired = await asyncio.wait_for(lock.acquire(), timeout)
+            if not acquired:
+                raise LockNotAcquiredException
+            yield acquired
+
+        except asyncio.TimeoutError:
             raise LockNotAcquiredException
-        yield acquired
 
     finally:
         if acquired:
             lock.release()
 
 
-class ADBPythonSync(object):
+class ADBPythonAsync(object):
     """A manager for ADB connections that uses a Python implementation of the ADB protocol.
 
     Parameters
@@ -75,13 +77,13 @@ class ADBPythonSync(object):
         self.host = host
         self.port = int(port)
         self.adbkey = adbkey
-        self._adb = AdbDeviceTcp(host=self.host, port=self.port, default_timeout_s=9.)
+        self._adb = AdbDeviceTcp(host=self.host, port=self.port, default_timeout_s=9., banner=b'aio-androidtv')
 
         # keep track of whether the ADB connection is intact
         self._available = False
 
         # use a lock to make sure that ADB commands don't overlap
-        self._adb_lock = threading.Lock()
+        self._adb_lock = asyncio.Lock()
 
     @property
     def available(self):
@@ -95,13 +97,13 @@ class ADBPythonSync(object):
         """
         return self._adb.available
 
-    def close(self):
+    async def close(self):
         """Close the ADB socket connection.
 
         """
-        self._adb.close()
+        await self._adb.close()
 
-    def connect(self, always_log_errors=True, auth_timeout_s=DEFAULT_AUTH_TIMEOUT_S):
+    async def connect(self, always_log_errors=True, auth_timeout_s=DEFAULT_AUTH_TIMEOUT_S):
         """Connect to an Android TV / Fire TV device.
 
         Parameters
@@ -118,7 +120,7 @@ class ADBPythonSync(object):
 
         """
         try:
-            with _acquire(self._adb_lock):
+            async with _acquire(self._adb_lock):
                 # Catch exceptions
                 try:
                     # Connect with authentication
@@ -136,11 +138,11 @@ class ADBPythonSync(object):
 
                         signer = PythonRSASigner(pub, priv)
 
-                        self._adb.connect(rsa_keys=[signer], auth_timeout_s=auth_timeout_s)
+                        await self._adb.connect(rsa_keys=[signer], auth_timeout_s=auth_timeout_s)
 
                     # Connect without authentication
                     else:
-                        self._adb.connect(auth_timeout_s=auth_timeout_s)
+                        await self._adb.connect(auth_timeout_s=auth_timeout_s)
 
                     # ADB connection successfully established
                     _LOGGER.debug("ADB connection to %s:%d successfully established", self.host, self.port)
@@ -154,7 +156,7 @@ class ADBPythonSync(object):
                         _LOGGER.warning("Couldn't connect to %s:%d.  %s: %s", self.host, self.port, exc.__class__.__name__, exc.strerror)
 
                     # ADB connection attempt failed
-                    self.close()
+                    await self.close()
                     self._available = False
                     return False
 
@@ -163,17 +165,17 @@ class ADBPythonSync(object):
                         _LOGGER.warning("Couldn't connect to %s:%d.  %s: %s", self.host, self.port, exc.__class__.__name__, exc)
 
                     # ADB connection attempt failed
-                    self.close()
+                    await self.close()
                     self._available = False
                     return False
 
         except LockNotAcquiredException:
             _LOGGER.warning("Couldn't connect to %s:%d because adb-shell lock not acquired.", self.host, self.port)
-            self.close()
+            await self.close()
             self._available = False
             return False
 
-    def pull(self, local_path, device_path):
+    async def pull(self, local_path, device_path):
         """Pull a file from the device using the Python ADB implementation.
 
         Parameters
@@ -188,12 +190,12 @@ class ADBPythonSync(object):
             _LOGGER.debug("ADB command not sent to %s:%d because adb-shell connection is not established: pull(%s, %s)", self.host, self.port, local_path, device_path)
             return
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via adb-shell: pull(%s, %s)", self.host, self.port, local_path, device_path)
-            self._adb.pull(device_path, local_path)
+            await self._adb.pull(device_path, local_path)
             return
 
-    def push(self, local_path, device_path):
+    async def push(self, local_path, device_path):
         """Push a file to the device using the Python ADB implementation.
 
         Parameters
@@ -208,12 +210,12 @@ class ADBPythonSync(object):
             _LOGGER.debug("ADB command not sent to %s:%d because adb-shell connection is not established: push(%s, %s)", self.host, self.port, local_path, device_path)
             return
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via adb-shell: push(%s, %s)", self.host, self.port, local_path, device_path)
-            self._adb.push(local_path, device_path)
+            await self._adb.push(local_path, device_path)
             return
 
-    def screencap(self):
+    async def screencap(self):
         """Take a screenshot using the Python ADB implementation.
 
         Returns
@@ -226,14 +228,14 @@ class ADBPythonSync(object):
             _LOGGER.debug("ADB screencap not taken from %s:%d because adb-shell connection is not established", self.host, self.port)
             return None
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Taking screencap from %s:%d via adb-shell", self.host, self.port)
-            result = self._adb.shell("screencap -p", decode=False)
+            result = await self._adb.shell("screencap -p", decode=False)
             if result[5:6] == b"\r":
                 return result.replace(b"\r\n", b"\n")
             return result
 
-    def shell(self, cmd):
+    async def shell(self, cmd):
         """Send an ADB command using the Python ADB implementation.
 
         Parameters
@@ -251,12 +253,12 @@ class ADBPythonSync(object):
             _LOGGER.debug("ADB command not sent to %s:%d because adb-shell connection is not established: %s", self.host, self.port, cmd)
             return None
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via adb-shell: %s", self.host, self.port, cmd)
-            return self._adb.shell(cmd)
+            return await self._adb.shell(cmd)
 
 
-class ADBServerSync(object):
+class ADBServerAsync(object):
     """A manager for ADB connections that uses an ADB server.
 
     Parameters
@@ -284,7 +286,7 @@ class ADBServerSync(object):
         self._was_available = False
 
         # use a lock to make sure that ADB commands don't overlap
-        self._adb_lock = threading.Lock()
+        self._adb_lock = asyncio.Lock()
 
     @property
     def available(self):
@@ -301,7 +303,7 @@ class ADBServerSync(object):
 
         return self._available
 
-    def close(self):
+    async def close(self):
         """Close the ADB server socket connection.
 
         Currently, this doesn't do anything except set ``self._available = False``.
@@ -309,7 +311,7 @@ class ADBServerSync(object):
         """
         self._available = False
 
-    def connect(self, always_log_errors=True):
+    async def connect(self, always_log_errors=True):
         """Connect to an Android TV / Fire TV device.
 
         Parameters
@@ -324,7 +326,7 @@ class ADBServerSync(object):
 
         """
         try:
-            with _acquire(self._adb_lock):
+            async with _acquire(self._adb_lock):
                 # Catch exceptions
                 try:
                     self._adb_client = Client(host=self.adb_server_ip, port=self.adb_server_port)
@@ -341,7 +343,7 @@ class ADBServerSync(object):
                     if self._was_available or always_log_errors:
                         _LOGGER.warning("Couldn't connect to %s:%d via ADB server %s:%d because the server is not connected to the device", self.host, self.port, self.adb_server_ip, self.adb_server_port)
 
-                    self.close()
+                    await self.close()
                     self._available = False
                     self._was_available = False
                     return False
@@ -351,19 +353,19 @@ class ADBServerSync(object):
                     if self._was_available or always_log_errors:
                         _LOGGER.warning("Couldn't connect to %s:%d via ADB server %s:%d, error: %s", self.host, self.port, self.adb_server_ip, self.adb_server_port, exc)
 
-                    self.close()
+                    await self.close()
                     self._available = False
                     self._was_available = False
                     return False
 
         except LockNotAcquiredException:
             _LOGGER.warning("Couldn't connect to %s:%d via ADB server %s:%d because pure-python-adb lock not acquired.", self.host, self.port, self.adb_server_ip, self.adb_server_port)
-            self.close()
+            await self.close()
             self._available = False
             self._was_available = False
             return False
 
-    def pull(self, local_path, device_path):
+    async def pull(self, local_path, device_path):
         """Pull a file from the device using an ADB server.
 
         Parameters
@@ -378,12 +380,12 @@ class ADBServerSync(object):
             _LOGGER.debug("ADB command not sent to %s:%d via ADB server %s:%d because pure-python-adb connection is not established: pull(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             return
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via ADB server %s:%d: pull(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             self._adb_device.pull(device_path, local_path)
             return
 
-    def push(self, local_path, device_path):
+    async def push(self, local_path, device_path):
         """Push a file to the device using an ADB server.
 
         Parameters
@@ -398,12 +400,12 @@ class ADBServerSync(object):
             _LOGGER.debug("ADB command not sent to %s:%d via ADB server %s:%d because pure-python-adb connection is not established: push(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             return
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via ADB server %s:%d: push(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             self._adb_device.push(local_path, device_path)
             return
 
-    def screencap(self):
+    async def screencap(self):
         """Take a screenshot using an ADB server.
 
         Returns
@@ -416,11 +418,11 @@ class ADBServerSync(object):
             _LOGGER.debug("ADB screencap not taken from %s:%d via ADB server %s:%d because pure-python-adb connection is not established", self.host, self.port, self.adb_server_ip, self.adb_server_port)
             return None
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Taking screencap from %s:%d via ADB server %s:%d", self.host, self.port, self.adb_server_ip, self.adb_server_port)
             return self._adb_device.screencap()
 
-    def shell(self, cmd):
+    async def shell(self, cmd):
         """Send an ADB command using an ADB server.
 
         Parameters
@@ -438,6 +440,6 @@ class ADBServerSync(object):
             _LOGGER.debug("ADB command not sent to %s:%d via ADB server %s:%d because pure-python-adb connection is not established: %s", self.host, self.port, self.adb_server_ip, self.adb_server_port, cmd)
             return None
 
-        with _acquire(self._adb_lock):
+        async with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via ADB server %s:%d: %s", self.host, self.port, self.adb_server_ip, self.adb_server_port, cmd)
             return self._adb_device.shell(cmd)
