+++ androidtv/adb_manager/adb_manager_sync.py
@@ -1,37 +1,40 @@
 """Classes to manage ADB connections.
 
 * :py:class:`ADBPython` utilizes a Python implementation of the ADB protocol.
+* :py:class:`ADBServer` utilizes an ADB server to communicate with the device.
 
 """
 
 
-import asyncio
-from contextlib import asynccontextmanager
+from contextlib import contextmanager
 import logging
+import sys
+import threading
 
-from aio_adb_shell.adb_device import AdbDeviceTcp
-from aio_adb_shell.auth.sign_pythonrsa import PythonRSASigner
+from adb_shell.adb_device import AdbDeviceTcp
+from adb_shell.auth.sign_pythonrsa import PythonRSASigner
 from ppadb.client import Client
 
-from .constants import DEFAULT_AUTH_TIMEOUT_S
-from .exceptions import LockNotAcquiredException
+from ..constants import DEFAULT_AUTH_TIMEOUT_S
+from ..exceptions import LockNotAcquiredException
 
 _LOGGER = logging.getLogger(__name__)
 
-#: Default timeout for acquiring the async lock that protects ADB commands
-DEFAULT_TIMEOUT = 3.0
+#: Use a timeout for the ADB threading lock if it is supported
+LOCK_KWARGS = {'timeout': 3} if sys.version_info[0] > 2 and sys.version_info[1] > 1 else {}
+
+if sys.version_info[0] == 2:  # pragma: no cover
+    FileNotFoundError = IOError  # pylint: disable=redefined-builtin
 
 
-@asynccontextmanager
-async def _acquire(lock, timeout=DEFAULT_TIMEOUT):
-    """Handle acquisition and release of an ``asyncio.Lock`` object with a timeout.
+@contextmanager
+def _acquire(lock):
+    """Handle acquisition and release of a ``threading.Lock`` object with ``LOCK_KWARGS`` keyword arguments.
 
     Parameters
     ----------
-    lock : asyncio.Lock
+    lock : threading.Lock
         The lock that we will try to acquire
-    timeout : float
-        The timeout in seconds
 
     Yields
     ------
@@ -45,22 +48,17 @@ async def _acquire(lock, timeout=DEFAULT_TIMEOUT):
 
     """
     try:
-        acquired = False
-        try:
-            acquired = await asyncio.wait_for(lock.acquire(), timeout)
-            if not acquired:
-                raise LockNotAcquiredException
-            yield acquired
-
-        except asyncio.TimeoutError:
+        acquired = lock.acquire(**LOCK_KWARGS)
+        if not acquired:
             raise LockNotAcquiredException
+        yield acquired
 
     finally:
         if acquired:
             lock.release()
 
 
-class ADBPythonAsync(object):
+class ADBPythonSync(object):
     """A manager for ADB connections that uses a Python implementation of the ADB protocol.
 
     Parameters
@@ -77,13 +75,13 @@ class ADBPythonAsync(object):
         self.host = host
         self.port = int(port)
         self.adbkey = adbkey
-        self._adb = AdbDeviceTcp(host=self.host, port=self.port, default_timeout_s=9., banner=b'aio-androidtv')
+        self._adb = AdbDeviceTcp(host=self.host, port=self.port, default_timeout_s=9.)
 
         # keep track of whether the ADB connection is intact
         self._available = False
 
         # use a lock to make sure that ADB commands don't overlap
-        self._adb_lock = asyncio.Lock()
+        self._adb_lock = threading.Lock()
 
     @property
     def available(self):
@@ -97,13 +95,13 @@ class ADBPythonAsync(object):
         """
         return self._adb.available
 
-    async def close(self):
+    def close(self):
         """Close the ADB socket connection.
 
         """
-        await self._adb.close()
+        self._adb.close()
 
-    async def connect(self, always_log_errors=True, auth_timeout_s=DEFAULT_AUTH_TIMEOUT_S):
+    def connect(self, always_log_errors=True, auth_timeout_s=DEFAULT_AUTH_TIMEOUT_S):
         """Connect to an Android TV / Fire TV device.
 
         Parameters
@@ -120,7 +118,7 @@ class ADBPythonAsync(object):
 
         """
         try:
-            async with _acquire(self._adb_lock):
+            with _acquire(self._adb_lock):
                 # Catch exceptions
                 try:
                     # Connect with authentication
@@ -138,11 +136,11 @@ class ADBPythonAsync(object):
 
                         signer = PythonRSASigner(pub, priv)
 
-                        await self._adb.connect(rsa_keys=[signer], auth_timeout_s=auth_timeout_s)
+                        self._adb.connect(rsa_keys=[signer], auth_timeout_s=auth_timeout_s)
 
                     # Connect without authentication
                     else:
-                        await self._adb.connect(auth_timeout_s=auth_timeout_s)
+                        self._adb.connect(auth_timeout_s=auth_timeout_s)
 
                     # ADB connection successfully established
                     _LOGGER.debug("ADB connection to %s:%d successfully established", self.host, self.port)
@@ -156,7 +154,7 @@ class ADBPythonAsync(object):
                         _LOGGER.warning("Couldn't connect to %s:%d.  %s: %s", self.host, self.port, exc.__class__.__name__, exc.strerror)
 
                     # ADB connection attempt failed
-                    await self.close()
+                    self.close()
                     self._available = False
                     return False
 
@@ -165,17 +163,17 @@ class ADBPythonAsync(object):
                         _LOGGER.warning("Couldn't connect to %s:%d.  %s: %s", self.host, self.port, exc.__class__.__name__, exc)
 
                     # ADB connection attempt failed
-                    await self.close()
+                    self.close()
                     self._available = False
                     return False
 
         except LockNotAcquiredException:
             _LOGGER.warning("Couldn't connect to %s:%d because adb-shell lock not acquired.", self.host, self.port)
-            await self.close()
+            self.close()
             self._available = False
             return False
 
-    async def pull(self, local_path, device_path):
+    def pull(self, local_path, device_path):
         """Pull a file from the device using the Python ADB implementation.
 
         Parameters
@@ -190,12 +188,12 @@ class ADBPythonAsync(object):
             _LOGGER.debug("ADB command not sent to %s:%d because adb-shell connection is not established: pull(%s, %s)", self.host, self.port, local_path, device_path)
             return
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via adb-shell: pull(%s, %s)", self.host, self.port, local_path, device_path)
-            await self._adb.pull(device_path, local_path)
+            self._adb.pull(device_path, local_path)
             return
 
-    async def push(self, local_path, device_path):
+    def push(self, local_path, device_path):
         """Push a file to the device using the Python ADB implementation.
 
         Parameters
@@ -210,12 +208,12 @@ class ADBPythonAsync(object):
             _LOGGER.debug("ADB command not sent to %s:%d because adb-shell connection is not established: push(%s, %s)", self.host, self.port, local_path, device_path)
             return
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via adb-shell: push(%s, %s)", self.host, self.port, local_path, device_path)
-            await self._adb.push(local_path, device_path)
+            self._adb.push(local_path, device_path)
             return
 
-    async def screencap(self):
+    def screencap(self):
         """Take a screenshot using the Python ADB implementation.
 
         Returns
@@ -228,14 +226,14 @@ class ADBPythonAsync(object):
             _LOGGER.debug("ADB screencap not taken from %s:%d because adb-shell connection is not established", self.host, self.port)
             return None
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Taking screencap from %s:%d via adb-shell", self.host, self.port)
-            result = await self._adb.shell("screencap -p", decode=False)
+            result = self._adb.shell("screencap -p", decode=False)
             if result[5:6] == b"\r":
                 return result.replace(b"\r\n", b"\n")
             return result
 
-    async def shell(self, cmd):
+    def shell(self, cmd):
         """Send an ADB command using the Python ADB implementation.
 
         Parameters
@@ -253,12 +251,12 @@ class ADBPythonAsync(object):
             _LOGGER.debug("ADB command not sent to %s:%d because adb-shell connection is not established: %s", self.host, self.port, cmd)
             return None
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via adb-shell: %s", self.host, self.port, cmd)
-            return await self._adb.shell(cmd)
+            return self._adb.shell(cmd)
 
 
-class ADBServerAsync(object):
+class ADBServerSync(object):
     """A manager for ADB connections that uses an ADB server.
 
     Parameters
@@ -286,7 +284,7 @@ class ADBServerAsync(object):
         self._was_available = False
 
         # use a lock to make sure that ADB commands don't overlap
-        self._adb_lock = asyncio.Lock()
+        self._adb_lock = threading.Lock()
 
     @property
     def available(self):
@@ -303,7 +301,7 @@ class ADBServerAsync(object):
 
         return self._available
 
-    async def close(self):
+    def close(self):
         """Close the ADB server socket connection.
 
         Currently, this doesn't do anything except set ``self._available = False``.
@@ -311,7 +309,7 @@ class ADBServerAsync(object):
         """
         self._available = False
 
-    async def connect(self, always_log_errors=True):
+    def connect(self, always_log_errors=True):
         """Connect to an Android TV / Fire TV device.
 
         Parameters
@@ -326,7 +324,7 @@ class ADBServerAsync(object):
 
         """
         try:
-            async with _acquire(self._adb_lock):
+            with _acquire(self._adb_lock):
                 # Catch exceptions
                 try:
                     self._adb_client = Client(host=self.adb_server_ip, port=self.adb_server_port)
@@ -343,7 +341,7 @@ class ADBServerAsync(object):
                     if self._was_available or always_log_errors:
                         _LOGGER.warning("Couldn't connect to %s:%d via ADB server %s:%d because the server is not connected to the device", self.host, self.port, self.adb_server_ip, self.adb_server_port)
 
-                    await self.close()
+                    self.close()
                     self._available = False
                     self._was_available = False
                     return False
@@ -353,19 +351,19 @@ class ADBServerAsync(object):
                     if self._was_available or always_log_errors:
                         _LOGGER.warning("Couldn't connect to %s:%d via ADB server %s:%d, error: %s", self.host, self.port, self.adb_server_ip, self.adb_server_port, exc)
 
-                    await self.close()
+                    self.close()
                     self._available = False
                     self._was_available = False
                     return False
 
         except LockNotAcquiredException:
             _LOGGER.warning("Couldn't connect to %s:%d via ADB server %s:%d because pure-python-adb lock not acquired.", self.host, self.port, self.adb_server_ip, self.adb_server_port)
-            await self.close()
+            self.close()
             self._available = False
             self._was_available = False
             return False
 
-    async def pull(self, local_path, device_path):
+    def pull(self, local_path, device_path):
         """Pull a file from the device using an ADB server.
 
         Parameters
@@ -380,12 +378,12 @@ class ADBServerAsync(object):
             _LOGGER.debug("ADB command not sent to %s:%d via ADB server %s:%d because pure-python-adb connection is not established: pull(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             return
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via ADB server %s:%d: pull(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             self._adb_device.pull(device_path, local_path)
             return
 
-    async def push(self, local_path, device_path):
+    def push(self, local_path, device_path):
         """Push a file to the device using an ADB server.
 
         Parameters
@@ -400,12 +398,12 @@ class ADBServerAsync(object):
             _LOGGER.debug("ADB command not sent to %s:%d via ADB server %s:%d because pure-python-adb connection is not established: push(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             return
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via ADB server %s:%d: push(%s, %s)", self.host, self.port, self.adb_server_ip, self.adb_server_port, local_path, device_path)
             self._adb_device.push(local_path, device_path)
             return
 
-    async def screencap(self):
+    def screencap(self):
         """Take a screenshot using an ADB server.
 
         Returns
@@ -418,11 +416,11 @@ class ADBServerAsync(object):
             _LOGGER.debug("ADB screencap not taken from %s:%d via ADB server %s:%d because pure-python-adb connection is not established", self.host, self.port, self.adb_server_ip, self.adb_server_port)
             return None
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Taking screencap from %s:%d via ADB server %s:%d", self.host, self.port, self.adb_server_ip, self.adb_server_port)
             return self._adb_device.screencap()
 
-    async def shell(self, cmd):
+    def shell(self, cmd):
         """Send an ADB command using an ADB server.
 
         Parameters
@@ -440,6 +438,6 @@ class ADBServerAsync(object):
             _LOGGER.debug("ADB command not sent to %s:%d via ADB server %s:%d because pure-python-adb connection is not established: %s", self.host, self.port, self.adb_server_ip, self.adb_server_port, cmd)
             return None
 
-        async with _acquire(self._adb_lock):
+        with _acquire(self._adb_lock):
             _LOGGER.debug("Sending command to %s:%d via ADB server %s:%d: %s", self.host, self.port, self.adb_server_ip, self.adb_server_port, cmd)
             return self._adb_device.shell(cmd)
