+++ androidtv/androidtv/androidtv_sync.py
@@ -13,7 +13,7 @@ from .. import constants
 _LOGGER = logging.getLogger(__name__)
 
 
-class AndroidTV(BaseTVSync, BaseAndroidTV):
+class AndroidTVSync(BaseTVSync, BaseAndroidTV):
     """Representation of an Android TV device.
 
     Parameters
